Fix: CalculatorItemQueue 수정
