--- Calculator/Calculator/Model/Queue.swift
@@ -1,36 +1,0 @@
-//
-//  Queue.swift
-//  Calculator
-//
-//  Created by 정재근 on 2022/09/06.
-//
-
-import Foundation
-
-struct Queue<T> {
-    private var queue: [T] = []
-    
-    public var count: Int {
-        return queue.count
-    }
-    
-    public var isEmpty: Bool {
-        return queue.isEmpty
-    }
-    
-    public mutating func enqueue(_ element: T) {
-        queue.append(element)
-    }
-    
-    public mutating func dequeue() -> T? {
-        return isEmpty ? nil : queue.removeFirst()
-    }
-    
-    public mutating func clearAll() {
-        queue = []
-    }
-    
-    public mutating func clearLast() {
-        queue.removeLast()
-    }
-}
